requirements.txt
getgauge
requests
pyjwt